take a u32 decoder output range

+++ crates/backend-uzu/src/encodable_block/classifier.rs
@@ -130,7 +130,7 @@ impl<B: Backend> Classifier<B> {
             .collect::<Box<[TrieNode]>>();
         let hidden = self
             .transformer
-            .encode(hidden, None, &BatchTopology::new(&nodes, true), Some(0..batch_dim as usize), None, None, encoder)
+            .encode(hidden, None, &BatchTopology::new(&nodes, true), Some(0..batch_dim), None, None, encoder)
             .map_err(ClassifierError::Backend)?
             .output
             .unwrap();
